repackage iDEA for PYPI
 * run.py replaced by 'idea-run' cmdline script
 * ViDEO.py replaced by 'idea-video' cmdline script
 * OPTiMAL_HYB.py replaced by 'idea-optimize' script
 * new 'idea-examples' script that starts jupyter
   notebook server in examples directory
 * documentation updated
