+++ iDEA/info.py
@@ -3,7 +3,7 @@
 # The short X.Y version.
 version = '0.1'
 # The full version, including alpha/beta/rc tags.
-release = '0.1.0'
+release = '0.1.1'
 
 def get_sha1():
     """Returns sha1 hash of last commit from git
